refactor(RankChart): logicを別ファイルに切り出し
