+++ src/components/RankChart/types.ts
@@ -1,0 +1,9 @@
+import type { ProcessedArticle } from "@/lib/article";
+
+export type Ranking = {
+  user: string;
+  articleCount: number;
+  rank: number;
+};
+
+export type Article = ProcessedArticle<true>;
